refactor: cc-0006 update icon
