+++ src/components/EditorSideBar/index.tsx
@@ -1,0 +1,52 @@
+import EditorSideBarItem from "./EditorSideBarItem";
+// import * as svg1 from "./src/assets/svg/blueprints.svg"
+import { Box, Flex } from "@chakra-ui/react";
+
+const EditorSideBar = () => {
+  const sideBarContent = [
+    {
+      sideBarItemText: "Blueprints",
+      // sideBarIconSrc: "/svg/blueprints.svg",
+      iconPath: "M17.48 6.44033L12.6 0.640333C12.4045 0.429129 12.1657 0.262614 11.8999 0.152161C11.6341 0.0417079 11.3476 -0.0100804 11.06 0.000333046H2C1.46957 0.000333046 0.960859 0.211047 0.585786 0.586119C0.210714 0.961192 0 1.4699 0 2.00033V22.0003C0 22.5308 0.210714 23.0395 0.585786 23.4145C0.960859 23.7896 1.46957 24.0003 2 24.0003H26C26.5304 24.0003 27.0391 23.7896 27.4142 23.4145C27.7893 23.0395 28 22.5308 28 22.0003V8.44033C28 7.9099 27.7893 7.40119 27.4142 7.02612C27.0391 6.65105 26.5304 6.44033 26 6.44033H17.48ZM3 2.94033H10.6L13.56 6.44033H3V2.94033ZM25 20.9403H3V9.44033H25V20.9403Z"
+    },
+    {
+      sideBarItemText: "Elements",
+      // sideBarIconSrc: "/svg/elements.svg",
+      iconPath: "M26 0H2C1.46957 0 0.960859 0.210714 0.585786 0.585786C0.210714 0.960859 0 1.46957 0 2V26C0 26.5304 0.210714 27.0391 0.585786 27.4142C0.960859 27.7893 1.46957 28 2 28H26C26.5304 28 27.0391 27.7893 27.4142 27.4142C27.7893 27.0391 28 26.5304 28 26V2C28 1.46957 27.7893 0.960859 27.4142 0.585786C27.0391 0.210714 26.5304 0 26 0ZM25 15.5H18.5V3H25V15.5ZM15.5 3V9.5H3V3H15.5ZM3 12.5H15.5V25H3V12.5ZM18.5 25V18.5H25V25H18.5Z"
+    },
+    {
+      sideBarItemText: "Estimator",
+      // sideBarIconSrc: "/svg//estimator.svg",
+      iconPath:"M10.6667 12.6H5.33333C4.62609 12.6 3.94781 12.305 3.44772 11.7799C2.94762 11.2548 2.66667 10.5426 2.66667 9.8C2.66667 9.05739 2.94762 8.3452 3.44772 7.8201C3.94781 7.295 4.62609 7 5.33333 7H12C12.3536 7 12.6928 7.1475 12.9428 7.41005C13.1929 7.6726 13.3333 8.0287 13.3333 8.4C13.3333 8.7713 13.4738 9.1274 13.7239 9.38995C13.9739 9.6525 14.313 9.8 14.6667 9.8C15.0203 9.8 15.3594 9.6525 15.6095 9.38995C15.8595 9.1274 16 8.7713 16 8.4C16 7.28609 15.5786 6.2178 14.8284 5.43015C14.0783 4.6425 13.0609 4.2 12 4.2H9.33333V1.4C9.33333 1.0287 9.19286 0.672601 8.94281 0.41005C8.69276 0.1475 8.35362 0 8 0C7.64638 0 7.30724 0.1475 7.05719 0.41005C6.80714 0.672601 6.66667 1.0287 6.66667 1.4V4.2H5.33333C3.91885 4.2 2.56229 4.79 1.5621 5.8402C0.561903 6.89041 0 8.31479 0 9.8C0 11.2852 0.561903 12.7096 1.5621 13.7598C2.56229 14.81 3.91885 15.4 5.33333 15.4H10.6667C11.3739 15.4 12.0522 15.695 12.5523 16.2201C13.0524 16.7452 13.3333 17.4574 13.3333 18.2C13.3333 18.9426 13.0524 19.6548 12.5523 20.1799C12.0522 20.705 11.3739 21 10.6667 21H4C3.64638 21 3.30724 20.8525 3.05719 20.59C2.80714 20.3274 2.66667 19.9713 2.66667 19.6C2.66667 19.2287 2.52619 18.8726 2.27614 18.6101C2.02609 18.3475 1.68696 18.2 1.33333 18.2C0.979711 18.2 0.640573 18.3475 0.390525 18.6101C0.140476 18.8726 0 19.2287 0 19.6C0 20.7139 0.421427 21.7822 1.17157 22.5698C1.92172 23.3575 2.93913 23.8 4 23.8H6.66667V26.6C6.66667 26.9713 6.80714 27.3274 7.05719 27.59C7.30724 27.8525 7.64638 28 8 28C8.35362 28 8.69276 27.8525 8.94281 27.59C9.19286 27.3274 9.33333 26.9713 9.33333 26.6V23.8H10.6667C12.0812 23.8 13.4377 23.21 14.4379 22.1598C15.4381 21.1096 16 19.6852 16 18.2C16 16.7148 15.4381 15.2904 14.4379 14.2402C13.4377 13.19 12.0812 12.6 10.6667 12.6Z"
+    },
+    {
+      sideBarItemText: "Preview",
+      // sideBarIconSrc: "/svg//preview.svg",
+      iconPath: "M14 3C18.5248 3 22.8894 7.78 24.6513 10C22.8894 12.22 18.5248 17 14 17C9.47521 17 5.1106 12.22 3.34874 10C5.1106 7.78 9.47521 3 14 3ZM14 0C7.73337 0 2.38772 6.16 0.385601 8.82C0.13504 9.16253 0 9.57577 0 10C0 10.4242 0.13504 10.8375 0.385601 11.18C2.38772 13.84 7.73337 20 14 20C20.2666 20 25.6123 13.84 27.6144 11.18C27.865 10.8375 28 10.4242 28 10C28 9.57577 27.865 9.16253 27.6144 8.82C25.6123 6.16 20.2666 0 14 0ZM14 5C13.2566 4.99672 12.5227 5.16799 11.8577 5.5C12.3053 5.52444 12.7353 5.68232 13.0922 5.95326C13.449 6.2242 13.7164 6.59576 13.8599 7.02C13.991 7.41631 14.0097 7.84114 13.9139 8.24741C13.8181 8.65369 13.6115 9.02552 13.3171 9.3217C13.0227 9.61787 12.6519 9.82681 12.2459 9.92536C11.8399 10.0239 11.4145 10.0082 11.0168 9.88C10.5922 9.73674 10.2202 9.46962 9.94897 9.11312C9.67775 8.75662 9.5197 8.32711 9.49524 7.88C9.16591 8.53827 8.99454 9.2641 8.99471 10C8.99471 10.9889 9.28826 11.9556 9.83825 12.7779C10.3882 13.6001 11.17 14.241 12.0846 14.6194C12.9992 14.9978 14.0056 15.0969 14.9765 14.9039C15.9474 14.711 16.8393 14.2348 17.5393 13.5355C18.2393 12.8363 18.716 11.9454 18.9091 10.9755C19.1022 10.0055 19.0031 9.00021 18.6243 8.08658C18.2454 7.17295 17.6039 6.39206 16.7808 5.84265C15.9577 5.29325 14.99 5 14 5Z"
+    },
+  ];
+  return (
+    <Box
+      bg="background.primary"
+      w="98px"
+      h="calc(100vh - 50px)"
+      position="fixed"
+      top="100px"
+      left="0"
+    >
+      <Flex align="center" justify="center" flexDirection="column" maxW="98px">
+        {sideBarContent.map((item) => (
+          <EditorSideBarItem
+            key={item.sideBarItemText}
+            item_text={item.sideBarItemText}
+            // icon_src={item.sideBarIconSrc}
+            icon_path={item.iconPath}
+    
+          ></EditorSideBarItem>
+        ))}
+      </Flex>
+    </Box>
+  );
+};
+
+export default EditorSideBar;
